refactor: Simplify and encapsulate confirmation prompts

+++ scripts/setup/nvm.sh
@@ -1,24 +1,29 @@
 #!/usr/bin/env bash
 
+source ./scripts/tools/clone.sh
+source ./scripts/tools/confirm.sh
+
+git_clone_nvm() {
+    git_clone_repo "https://github.com/nvm-sh/nvm.git" "${HOME}/.nvm"
+}
+
 setup_nvm() {
+    confirm_proceed "nvm: Install and setup Node Version Manager" || return
+
     # Check if nvm is already installed
     if command -v nvm &> /dev/null; then
         echo "nvm is already installed"
-    else
-        # Download and install nvm
-        if ! git clone https://github.com/nvm-sh/nvm.git ~/.nvm; then
-            echo "Failed to install nvm to ${HOME}/.nvm"
-            exit 1
-        fi
-
-        # Source nvm script
-        source ~/.nvm/nvm.sh
+    else  # Download and install nvm
+        git_clone_nvm || { echo "Failed to install nvm to ${HOME}/.nvm"; return 1; }
     fi
+
+    # Source nvm script
+    source "${HOME}/.nvm/nvm.sh"
 
     # Install Node.js LTS version
     if ! nvm install --lts; then
         echo "Failed to install Node.js"
-        exit 1
+        return 1
     fi
 
     # Use installed Node.js version
@@ -27,7 +32,7 @@ setup_nvm() {
     
     if [ -z "$node_version" ]; then
         echo "Failed to find installed Node.js using ${node_version}"
-        exit 1
+        return 1
     fi
 
     nvm use "$node_version"
